move hint requested logic from container to action
