+++ src/views/buttons/HintButton.js
@@ -2,7 +2,8 @@ import React from "react";
 import { PropTypes } from "prop-types";
 
 export function HintButton(props) {
-  //since set code is async, i can mash the hint button and break it in timed mode
+  //since set code is async, user can hit the the hint button repeatly and break it in timed mode
+  //will only display when correctCode is set.
   let { requestHint, hintCount, correctCode } = props;
   return correctCode && hintCount < correctCode.code.length - 2 ? (
     <button onClick={() => requestHint()}>Request Hint</button>
